@@ -9,9 +9,7 @@
     <version>1.0-SNAPSHOT</version>
 
     <properties>
-        <maven.compiler.source>17</maven.compiler.source>
-        <maven.compiler.target>17</maven.compiler.target>
-        <java.version>17</java.version>
+        <java.version>21</java.version>
         <lombok.version>1.18.26</lombok.version>
         <commons-text.version>1.9</commons-text.version>
         <logback-classic.version>1.2.11</logback-classic.version>
