@@ -2,6 +2,7 @@ package pdx.fo76.asasm.instruction;
 
 import lombok.RequiredArgsConstructor;
 import pdx.fo76.asasm.SyntaxConstants;
+import pdx.fo76.asasm.util.ParseUtil;
 
 @RequiredArgsConstructor
 public class TypeName implements Identifier {
@@ -14,11 +15,11 @@ public class TypeName implements Identifier {
     }
 
     public static Identifier parse(String str) {
-        var prefix = str.substring(0, str.indexOf("("));
-        if (!prefix.equals(SyntaxConstants.TYPENAME)) {
+        var prefix = ParseUtil.callSiteName(str);
+        if (!SyntaxConstants.TYPENAME.equals(prefix)) {
             throw new IllegalArgumentException();
         }
-        var args = str.substring(str.indexOf("(") + 1, str.lastIndexOf(")"));
+        var args = ParseUtil.stripParentheses(str);
         var qns = args.substring(0, args.indexOf("<"));
         var param = args.substring(args.indexOf("<") + 1, args.lastIndexOf(">"));
         var qn = QName.parse(qns);
