@@ -1,8 +1,4 @@
 package pdx.fo76.asasm.syntax;
 
-import lombok.Value;
-
-@Value
-public class ParseStop {
-    int position;
+public record ParseStop(int position) {
 }
