@@ -1,11 +1,4 @@
 package pdx.fo76.asasm.token;
 
-import lombok.Value;
-
-@Value
-public class ParsedToken {
-    Token token;
-    String value;
-    int line;
-    int position;
+public record ParsedToken(Token token, String value, int line, int position) {
 }
